Updates to manifest and optimized mode
This commit makes updates to optimized mode. The current optimize mode
now refactors N->1 IO to N->m where m is the number of nodes. All
processes on a node reserve and write a stripe of the local file. The
mode attempts to make sure the load is spread out over a couple of
osts to avoid locking issues. Support for group locking is also
included in this commit but it appears to be disabled on LANL Lustre.

There is currently a scaleabilty issue with the manifest gather. It is
best to turn off bzip2 support for large runs until this issue is
resolved.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>